Change app.listen to work with Heroku
http://stackoverflow.com/questions/15693192/heroku-node-js-error-web-pro
cess-failed-to-bind-to-port-within-60-seconds-of

Heroku dynamically assigns your app a port, so you can't set the port
to a fixed number. Heroku adds the port to the env, so you can pull it
from there. Switch your listen to this:

.listen(process.env.PORT || 5000)

That way it'll still listen to port 5000 when you test locally, but it
will also work on Heroku.

You can check out the Heroku docs on Node.js here.